Move parser into its own subdirectory
