+++ arangod/Aql/Parser/tokens.ll
@@ -37,9 +37,9 @@ class Parser;
 
 #include "Aql/Ast.h"
 #include "Aql/AstNode.h"
-#include "Aql/grammar.hpp"
+#include "Aql/Parser/grammar.hpp"
 #include "Aql/Functions.h"
-#include "Aql/Parser.h"
+#include "Aql/Parser/Parser.h"
 #include "Aql/QueryContext.h"
 
 #define YY_EXTRA_TYPE arangodb::aql::Parser*
