+++ arangod/Aql/StandaloneCalculation.cpp
@@ -31,7 +31,7 @@
 #include "Aql/LazyConditions.h"
 #include "Aql/Optimizer.h"
 #include "Aql/OptimizerRule.h"
-#include "Aql/Parser.h"
+#include "Aql/Parser/Parser.h"
 #include "Aql/QueryContext.h"
 #include "Aql/QueryString.h"
 #include "Basics/Exceptions.h"
