+++ arangod/IResearch/IResearchAqlAnalyzer.cpp
@@ -34,7 +34,7 @@
 #include "Aql/FixedVarExpressionContext.h"
 #include "Aql/Optimizer.h"
 #include "Aql/OptimizerRule.h"
-#include "Aql/Parser.h"
+#include "Aql/Parser/Parser.h"
 #include "Aql/QueryContext.h"
 #include "Aql/QueryString.h"
 #include "Aql/SharedQueryState.h"
